+++ bcap_core/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package bcap_core
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.3 (2019-09-23)
+------------------
+
 3.0.2 (2017-12-15)
 ------------------
 * Change descriptions and add url, author