+++ bcap_service/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package bcap_service
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.3 (2019-09-23)
+------------------
+* Merge pull request `#11 <https://github.com/DENSORobot/denso_robot_ros/pull/11>`_
+
 3.0.2 (2017-12-15)
 ------------------
 * Remove new line from ROS_INFO,ROS_WARN,ROS_ERROR