+++ denso_robot_bringup/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package denso_robot_bringup
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.3 (2019-09-23)
+------------------
+* Change update_joint_limits.py `#23 <https://github.com/DENSORobot/denso_robot_ros/issues/23>`_
+* Add joint_state_publisher
+
 3.0.2 (2017-12-15)
 ------------------
 * Change descriptions and add url, author