+++ denso_robot_core/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package denso_robot_core
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.3 (2019-09-23)
+------------------
+* Change bCap service UDP timeout
+
 3.0.2 (2017-12-15)
 ------------------
 * Add I/O function for denso_robot_core