+++ denso_robot_gazebo/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package denso_robot_gazebo
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.3 (2019-09-23)
+------------------
+
 3.0.2 (2017-12-15)
 ------------------
 * Change descriptions and add url, author