+++ denso_robot_moveit_config/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package denso_robot_moveit_config
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.3 (2019-09-23)
+------------------
+* Fix joint_limits.yaml format
+
 3.0.2 (2017-12-15)
 ------------------
 * update version to 3.0.2